Move Dockerfile to seperate repo
To avoid having a new Docker container be built with every push to the
thesis repo the Dockerfile has been moved to a new repo (LaTeX Docker)
(https://github.com/matthewfeickert/LaTeX-Docker) and then the resulting
image is used. The Dockerfile that existed in the thesis repo can then
be removed.

Add "StrictHostKeyChecking no" to ssh config to avoid error in the CI