Ignore testing branches
In the Circle CI ignore any branch whose name starts with "test-". This
is to ensure that debugging doesn't launch CI jobs, but that the
`master` branch and any branches where chapters are being built out are
still tested. The choice of `ignore` over `only` is motivated by wanting
to be able to have a branch with name starting with "chapter-" or
"section-" that will still be tested with changes pushed. Thus, a PR of
a chapter can be worked on and still have the changes be broadcast to
collaborators.