Apply ssh config to all hosts
Given the configuration of Docker being used and the lack of ssh keys
the ssh config now applies "StrictHostKeyChecking no" to all hosts.
While it is perhaps safer to apply it to each individual host name, this
would need to be applied for each host name regardless, and so this
saves lines.