Add Travis CI
This CI is different than Circle CI in the sense that instead of
immediatley going into the Docker container the Docker container is
pulled into the current environment. Thus, Docker commands like run, cp,
exec must be executed.

The configuration is much slower than Circle CI so for the time being it
will only be run on the master branch. As a result, Circle CI will not
be run on master, but will be used for all other branches of
development.

Travis CI also allows the ability to specify different emails to
contact, so in the future members of the Ph.D. thesis committee can be
emailed alerts when builds have finished and drafts deployed.