Add ATLAS detector terms to glossary
Add:
- Inner Detector systems
- muon system terms
- TDAQ and HLT
- RoI

Also format the glossary to be programatically generated

As 'text' is the field that is displayed in the document text and 'name'
is the field that is displayed in the glossary along with the definition
make this clear my putting text as the first field. Additionally if a
term has an acronym or initialization give it in the name.

Have the glossary fields be defined in reference to the entry text so
that generation is as easy as possible.

This style of glossary does invite the idea though that the glossary
should all more of a listing of acronyms and not short definitions. This
should be revisited and cleared up.